modify(#167):previewPage modal img 정상화 시키는 김에 로직 분리 시키면 좋을거 같아서 분리했음
